Fix some tricky special cases with vector extension methods and null coalescing
Essentially, if we did something like:

    declare v: Vector2 | null
    v.mul

We'd compile it into

    funcref(v, "mul")

Which is obviously wrong because there is no "mul" on v, so we need to actually do

    funcref(self, "mul_vec_lib")

We can fix that easily, but what if we now do:

    declare v: Vector2 | null
    v?.mul(5)

We'd compile that into

    var __gen = [funcref(self, "mul_vec_lib"), {}]
    __gen.call_func(5)

This is once again wrong because we need to pass in v to mul_vec_lib.
Unfortuantely by the time that we call_func, we don't have a reference to v any
more, and there's no easy way to traverse the AST to get it, because someone could have done

    const foo = v?.mul

and then done other work before calling foo. The only way to solve this is to keep a reference to v
in the closure object we pass around for functions, e.g.

var __gen = [funcref(self, "mul_vec_lib"), {}, v]

Now we can call mul_vec_lib appropriately.